[Graphics] Use XRender and pixmaps
Instead of using imlib2 utilities to display images, upload them as
pixmaps and use XRender to blend them onto the window.

This should be much faster, because imlib2 does something like reading
a portion of the window from the X server, blending it on the client,
and then uploading it every single time an image is displayed.

Expect some bugs and compatibility issues.